# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Authentication and User Management API",
    "version": "1.0.0",
    "description": "This API provides endpoints for managing user authentication, authorization, and profile operations. It includes features like user registration, login, two-factor authentication, password and email updates, and various system-level functionalities such as retrieving hooks and states. Designed to support secure and scalable authentication workflows."
  },
  "servers": [
    {
      "url": "http://localhost:80"
    }
  ],
...
  "paths": {
...
    "/api/authentication/registration": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Create a new user",
        "description": "Register a new user and send a confirmation email to the provided email address.",
        "parameters": [
          {
            "name": "x-xsrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "XSRF token obtained from the `/api/system/xsfr` endpoint"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "The username of the new user",
                    "example": "johndoe"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "The email address of the new user",
                    "example": "johndoe@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "The password for the new user",
                    "example": "mypassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered"
          },
          "400": {
            "description": "Invalid input data"
          },
          "403": {
            "description": "Admin isn't created yet"
          },
          "409": {
            "description": "User with the same email address does already exist"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/authentication/signin": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Sign in",
        "description": "Authenticate a user with their email and password.",
        "parameters": [
          {
            "name": "x-xsrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "XSRF token obtained from the `/api/system/xsfr` endpoint"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "The email address of the user",
                    "example": "johndoe@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "The password for the user",
                    "example": "mypassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully signed in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "The authentication token. It is only present if two-factor authentication is not enabled.",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    },
                    "hook": {
                      "type": "string",
                      "nullable": true,
                      "description": "A hook set by the admin that can be used to redirect the user after a successful authentication.",
                      "example": "http://example.com/dashboard"
                    },
                    "twoFactorAuthId": {
                      "type": "string",
                      "nullable": true,
                      "description": "If two-factor authentication is enabled, this field will be populated with an identifier for the two-factor authentication process.",
                      "example": "5f8d0d55b54764421b7156fe"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/authentication/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logout",
        "description": "Logs out the authenticated user.",
        "parameters": [
          {
            "name": "x-xsrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "XSRF token obtained from the `/api/system/xsfr` endpoint"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully logged out"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/authentication/reset/password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Request password reset",
        "description": "Initiates the password reset process by sending a reset password email to the user's email address.",
        "parameters": [
          {
            "name": "x-xsrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "XSRF token obtained from the `/api/system/xsfr` endpoint"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "The email address of the user requesting a password reset.",
                    "example": "johndoe@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset request processed successfully. If the user exists, an email is sent."
          },
          "400": {
            "description": "Invalid input data, e.g., missing email."
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/authentication/token": {
      "put": {
        "tags": ["Authentication"],
        "summary": "Refresh authentication token",
        "description": "Creates a new authentication token and sets a new refresh token as an HttpOnly cookie. Requires a valid refresh token sent as a cookie.",
        "parameters": [
          {
            "name": "x-xsrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "XSRF token obtained from the `/api/system/xsfr` endpoint"
          }
        ],
        "responses": {
          "200": {
            "description": "Authentication token successfully refreshed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "The new authentication token.",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "description": "Contains the new refresh token.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Refresh token is missing or invalid."
          },
          "404": {
            "description": "User not found for the provided refresh token."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
...
  }
}